refactor: Relocate computeRealizedLpFeePct test
 - Factor the computeRealizedLpFeePct test out to a separate file, since
   it's directly related to the HubPoolClient and only indirectly
   related to the ConfigStoreClient.
 - Remove a lot of init cruft from the ConfigStoreClient test. As it
   turns out, this  was only necessary for the computeRealizedLpFeePct
   test.

As background, there have historically been ghosts in the
ConfigStoreClient test, as evidenced by the existing need to reset the
hre state prior to commencing the test. It seems possible to drop that
bodge now, so this consolidation _might_ actually be hinting at the
underlying issue.
